+++ src/TaskEither.ts
@@ -220,9 +220,19 @@ export const toUnion =
  * @category combinators
  * @since 2.0.0
  */
-export const orElse: <E, A, M>(onLeft: (e: E) => TaskEither<M, A>) => (ma: TaskEither<E, A>) => TaskEither<M, A> =
+export const orElse: <E1, A, E2>(onLeft: (e: E1) => TaskEither<E2, A>) => (ma: TaskEither<E1, A>) => TaskEither<E2, A> =
   /*#__PURE__*/
   ET.orElse_(T.Monad)
+
+/**
+ * Less strict version of [`orElse`](#orElse).
+ *
+ * @category combinators
+ * @since 2.10.0
+ */
+export const orElseW: <E1, E2, B>(
+  onLeft: (e: E1) => TaskEither<E2, B>
+) => <A>(ma: TaskEither<E1, A>) => TaskEither<E2, A | B> = orElse as any
 
 /**
  * @category combinators